+++ hibike/devices/Potentiometer/potentiometer.cpp
@@ -42,7 +42,7 @@ uint32_t device_write(uint8_t param, uint8_t* data, size_t len){
 //
 // You can use the helper function append_buf.
 // append_buf copies the specified amount data into the dst buffer and increments the offset
-uint8_t device_data_update(int param, uint8_t* data_update_buf, size_t buf_len) {
+uint8_t device_data_update(uint8_t param, uint8_t* data_update_buf, size_t buf_len) {
   // Read sensor, which is in uint16
   if (MAX_PAYLOAD_SIZE - buf_len < sizeof(float) || param>2 || param<0) {
     return 0;
